buffer_cache/requirements.txt
pyyaml>=6.0
openai>=1.0        # optional: for LLM-guided mutation + Python-to-C translation
anthropic>=0.20    # optional: alternative LLM provider
